Changed PETSc module filename to parafem_petsc.F90 to match the module name.
Restructured the makefiles for parafem_petsc.  The ParaFEM MPI
interface (mp_interface.f90) uses mpif.h rather than the MPI
module, so there are conflicts.  Now the PETSc library is
set up using include files instead of modules, and so the
parafem_petsc compilation depends on the ParaFEM mpi module
compilation.  The dependencies are not perfect but work.

Solvers are now chosen on the command line and a second (optional)
argument.  Helper functions are in the choose_solvers module.

The PETSc interface is still by direct calls and the .petsc file
is used again rather than the suggested .solvers file.