Umats now in xx15 and not large_strain.
Umat name specified on last line of control file (.dat) and
a new routine read_data_xx15 provided in xx15.F90

Umat choice is still a simple if, else if sequence - better to
have a umat moduel and a procedure pointer.

Still with the trunk xx15, plasticity and the old Souza necking umat.
This all has to be made consistent.

Compile and runs but the results for PETSc GMRES/Jacobi are now
different.