It would be possible to use the un-rearranged eigenprojections but that would need a change in LN_DERIV which would need to be tested.
The tests for eigenvalue equality missed the case where the difference
was equal to the tolerance.  Note that the tolerance is OK for double
precision, and OK to use directly because the eigenvalues are near 1.

The tests in EIGEN and LN_DERIV should be identical. To make them
identical the scaled calculation has been replaced with an unscaled
calculation in EIGEN.

There is now EIGEN_SCALED (the old subroutine but with corrected tests
for eigenvalue equality), EIGEN (the unscaled and corrected ParaFEM
method) EIGEN_LAPACK, and EIGEN_KOPP.  Tests using one_element_compression
show that EIGEN_SCALED, EIGEN, and EIGEN_LAPACK match the original results
for the starin but now EIGEN_KOPP does not and the simulation fails.