Working towards merging the petsc branch with trunk.
Default build is without PETSc.  The install step now only installs
the modules that have been built (is this correct?).  There is no
need for sleeps in the build script now that the makefiles have been
corrected (&s changed to &&s).

(The ARCHER TDS modules are now working, so the workaround has been
removed.)